--- otr/Assets/CopCarAudio.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: de12559c0b9d7a1498af56438950f3f6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
