--- otr/Assets/GeneralObjectCollisionDetector.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5dc448d0e0ae0054585de3fa76d74466
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
