--- otr/Assets/MainMenu.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 141a720efd53f42bbba061383128bcb6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
