+++ otr/Assets/Scenes/GameScene.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 03a9decd1be974abea547822bd3a7396
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
