+++ otr/Assets/Scenes/GameScene/NavMesh.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ff7664db3099146c39fd1f1876c04cdd
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 23800000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
