+++ otr/Assets/Scripts/Camera/CameraFollow.cs
@@ -4,29 +4,6 @@ using UnityEngine;
 
 public class CameraFollow : MonoBehaviour
 {
-    // [SerializeField] private Vector3 offset;
-    // [SerializeField] private Transform target;
-    // [SerializeField] private float translateSpeed;
-    // [SerializeField] private float rotationSpeed;
-
-    // void FixedUpdate()
-    // {
-    //     HandleTranslation();
-    //     HandleRotation();
-    // }
-
-    // private void HandleTranslation()
-    // {
-    //     var targetPosition = target.TransformPoint(offset);
-    //     transform.position = Vector3.Lerp(transform.position, targetPosition, translateSpeed * Time.deltaTime);
-    // }
-
-    // private void HandleRotation()
-    // {
-    //     var direction = target.position - transform.position;
-    //     var rotation = Quaternion.LookRotation(direction, Vector3.up);
-    //     transform.rotation = Quaternion.Lerp(transform.rotation, rotation, rotationSpeed * Time.deltaTime);
-    // }
 
     public GameObject Player;
     public GameObject child;
@@ -37,6 +14,17 @@ public class CameraFollow : MonoBehaviour
         Player = GameObject.FindGameObjectWithTag("Player");
         child = Player.transform.Find("camera constraint").gameObject;
 
+    }
+
+    public void Update() {
+        if (Input.GetKey(KeyCode.RightShift))
+        {
+            child = Player.transform.Find("camera constraint (reverse)").gameObject;
+        }
+        else 
+        {
+            child = Player.transform.Find("camera constraint").gameObject;
+        }
     }
 
     private void FixedUpdate()
