+++ otr/Assets/Scripts/Cars/PlayerCarController.cs
@@ -91,7 +91,6 @@ public class PlayerCarController : MonoBehaviour
         horizontalInput = Input.GetAxisRaw(HORIZONTAL);
         verticalInput = Input.GetAxisRaw(VERTICAL);
         isBraking = Input.GetKey(KeyCode.Space);
-        print(isBraking);
     }
 
     private void UpdateWheels()
