+++ otr/Assets/Scripts/CopFollow.cs
@@ -1,0 +1,28 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.AI;
+
+public class CopFollow : MonoBehaviour
+{
+    private NavMeshAgent Cop;
+    public GameObject Player;
+    public float MobDistanceRun = 4.0f;
+
+    // Start is called before the first frame update
+    void Start()
+    {
+        Cop = GetComponent<NavMeshAgent>();
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        float distance = Vector3.Distance(transform.position, Player.transform.position);
+
+        Vector3 dirToPlayer = transform.position - Player.transform.position;
+        Vector3 newPos = transform.position - dirToPlayer;
+
+        Cop.SetDestination(newPos);
+    }
+}
