+++ otr/Assets/Scripts/CopFollow.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 82315a1907beb4281b1ce760537a9fb9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
