+++ otr/Assets/Scripts/GameOverScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d66bd0dbca351f7479d5b6ea5c43f8b9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
