+++ otr/Assets/Scripts/MainMenu.cs
@@ -1,0 +1,16 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+public class MainMenu : MonoBehaviour
+{
+    // Start is called before the first frame update
+    public void PlayGame() {
+        SceneManager.LoadScene(1);
+    }
+
+    public void QuitGame() {
+        Application.Quit();
+    }
+}
